+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ## 0.0.1
 
 - Initial release.
-- Added `TextWithContoursPack` widget.
-
+- Added `TextWithOuline` widget.